+++ scripts/signup.js
@@ -8,6 +8,7 @@ import { checkPassword } from "./input/checkPassword.js";
 import { inputEmail, inputPassword } from "./constants.js";
 import { emailDuplicationCheck } from "./input/emailDuplicationCheck.js";
 import { checkEmailFormat } from "./input/checkEmailFormat.js";
+import { postNewAccount } from "./accounts/postNewAccount.js";
 
 const eyeImagePasswordEl = document.querySelector("#eyeImage-password");
 const eyeImagePasswordReEl = document.querySelector("#eyeImage-password-re");
@@ -38,7 +39,7 @@ inputEmail.addEventListener("focusout", checkEmailFormat);
 //<비밀번호 형식 검증>
 inputPassword.addEventListener("focusout", checkPassword);
 
-//비밀번호-비밀번호 확인 값 일치 확인
+//<비밀번호-비밀번호 확인 값 일치 확인>
 function checkPasswordRe() {
   const password = inputPassword.value;
   const passwordRe = inputPasswordRe.value;
@@ -55,53 +56,29 @@ function checkPasswordRe() {
 }
 inputPasswordRe.addEventListener("focusout", checkPasswordRe);
 
-//이메일&비밀번호 모두 유효한 값이라면 버튼 클릭시 /folder로 이동하기
-function submitAccount() {
+//<이메일&비밀번호 모두 유효한 값이라면 버튼 클릭시 /folder로 이동하기>
+async function submitAccount() {
   const isEmailValid = checkEmailFormat();
   const isPasswordValid = checkPassword();
   const isPasswordReValid = checkPasswordRe();
+  const isEmailDuplicationCheck = await emailDuplicationCheck(inputEmail.value);
 
-  if (isEmailValid && isPasswordValid && isPasswordReValid) {
-    async function newAccount(email, password) {
-      //POST로 새로운 계정정보를 추가
-      const newAccountInfo = {
-        email: email,
-        password: password,
-      };
-
-      try {
-        const response = await fetch(
-          "https://bootcamp-api.codeit.kr/api/sign-up",
-          {
-            method: "POST",
-            headers: {
-              "Content-Type": "application/json",
-            },
-            body: JSON.stringify(newAccountInfo),
-          }
-        );
-
-        if (response.status === 200) {
-          const responseData = await response.json();
-          const accessToken = responseData.accessToken;
-          window.localStorage.setItem("accessToken", accessToken);
-          window.location.href = "/folder";
-        }
-      } catch (error) {
-        console.error(error);
-      }
-    }
+  if (
+    isEmailValid &&
+    isPasswordValid &&
+    isPasswordReValid &&
+    isEmailDuplicationCheck
+  ) {
+    postNewAccount(inputEmail.value, inputPassword.value);
   }
 }
 
 submitButton.addEventListener("click", function (event) {
   event.preventDefault();
-  emailDuplicationCheck();
   submitAccount();
 });
 
 signupForm.addEventListener("submit", function (event) {
   event.preventDefault();
-  emailDuplicationCheck();
   submitAccount();
 });
